Move a uncommon method to sub-class
