+++ lib/convergence/command.rb
@@ -32,16 +32,6 @@ class Convergence::Command
                 end
   end
 
-  def sql_generator
-    @sql_generator ||= case database_adapter
-                       when 'mysql', 'mysql2'
-                         require 'convergence/sql_generator/mysql_generator'
-                         SQLGenerator::MysqlGenerator.new
-                       else
-                         fail NotImplementedError.new('unknown database adapter')
-                       end
-  end
-
   def logger
     @logger ||= Convergence::Logger.new
   end
